pyproject.toml
[tool.poetry]
name = "employee-management"
version = "0.1.0"
description = ""
authors = ["mercybassey <udohmercy911@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
python-dotenv = "^1.0.1"
permit = "^2.7.1"
authlib = "^1.4.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
